v0.5.2
No new features. Added substantial type annotation helpers, real-time data checking and error reporting, and misc bug fixes.

- Added test_actions.py to test default Lie group actions from Python.
- Added assertion handler "lielab_assert" for handling real time reporting of errors.
- Added annotations to many objects in Python.
- Added dataproxy to CompositeAlgebra, Group, and Manifold.
- Added iterator support to Composite objects.
- Added many list-like operations in Python for Composite objects.
- Added many assertions for major runtime problems for various integration methods.
- Added set item capability to operator[] for Composite types.
- Added Lielab::common target for common options shared across each static lib.
- Added sinc function.
- Added check_almost_equal_tol to testing.
- Added check_almost_equal_nulp to testing.
- Added initialization of SE by (RN, SO).
- Added initialization of se with (rn, so).
- Added topology checks to solve_ivp.
- Added CrouchGrossman methods to IVP methods.
- Added get_simd_info(), get_eigen_info(), and get_pybind11_info() functions that provide info when Lielab is called as a pre-compiled library.
- Added pickling support to all objects in domain as well as IVPSolution.
- Added rebase_every_step as an IVPOption to toggle MuntheKaas solution rebase around 0 every time step, or to let the solution drift.
- Added vendored in licenses to python wheels.
- Changed IVPSettings to IVPCommon.
- Changed GrR to full name Grassmannian.
- Changed rootsolver for IVP event crossing from Newton to Golden for robustness.
- Changed behavior with SO(1) to have size 0.
- Changed eigentools inputs use int instead of ptrdiff_t.
- Changed eigentools uses Eigen::Index instead of ptrdiff_t.
- Changed behavior of int initialization of Composite objects to initialize as n-empty objects instead of shape-n.
- Changed from_shape() static initializer for Lie groups to identity().
- Changed from_shape() static initializer for Lie algebras to zero().
- Changed order type in exponential from size_t to int to better handle user input issues.
- Changed SE's data to be (RN, SO) to better represent the semidirect product.
- Changed se to use (rn, so) as a data type to properly model the semidirect product as opposed to the monolithic matrix. This enables reuse of rn and so operations.
- Fixed a bug in scaling of RungeKutta and MuntheKaas methods where the absolute value and maximum operations weren't taken in the correct order.
- Fixed bug in adjoint for se(3) and se(4) that returned the wrong shape.
- Fixed bug in dexpinv for se where dexp_numerical was called instead of dexpinv_numerical.
- Improved real time error handling of various functions.
- Improved IVPSolution data trimming by resizing instead of copying.
- Improved performance in eigentools by reducing some unnecessary copy operations.
- Improved multiplication and inverse operations of SE by the semidirect product structure.
- Modified coefficients of dexp and dexpinv for so(3) to use Lagrange polynomial approximation near the singularity at 0.
- Moved some documentation to more generic markdown files.
- Moved and organized some pre-existing optimization routines to "optimize" submodule.
- Raising Catch2 version to 3.11.0.
- Removed Python 3.8 and 3.9 support.
- Removed unused Lie algebra actions.
- Removed assert_matrix in favor of check_almost_equal_X.
- Removed assert_domain in favor of check_topology.
- Removed is_algebra.
- Removed many hardcoded type checking in favor of visit in various functions.
- Removed a substantial amount of manual logic in Composite objects by using visit.
- Removed some complex number support from su to prevent out of algebra operations.
- Removed get_matrix() method from CompositeManifold.
- Removed from_shape() initialization from CompositeManifold.
- Removed get_shape(s)() method from CompositeManifold.
- Removed operator() from CompositeManifold.
- Renamed data to point for clarity of purpose.
- Removed factorial in favor of built in tgamma.
- Removed matstr.
- Removed operator<< from domain.
- Removed operator= from domain objects.
- Updated se/SE projections to reuse so/SO projections.
- Using operator() for general Lie group/algebra IO operations and operator[] for manifold specific IO operations.
- Using std::visit where possible for Composite objects.

+++ Lielab/functions/Killing.tpp
@@ -1,7 +1,6 @@
 #ifndef LIELAB_FUNCTIONS_KILLING_TPP
 #define LIELAB_FUNCTIONS_KILLING_TPP
 
-#include "commutator.hpp"
 #include "Killing.hpp"
 
 #include "Lielab/domain.hpp"
@@ -19,24 +18,24 @@ double Killing(const LA& a, const LA& b)
     * @param[out] k Killing coefficient between a and b.
     */
 
-    const size_t shape = a.get_shape();
+    const int shape = a.get_shape();
 
     if (shape != b.get_shape())
     {
         throw Lielab::utils::InputError("Killing: Shapes of a and b must be equal.");
     }
 
-    const size_t dim = a.get_dimension();
+    const int dim = a.get_dimension();
 
     std::vector<LA> basis = std::vector<LA>(dim);
 
-    for (size_t ii = 0; ii < dim; ii++)
+    for (int ii = 0; ii < dim; ii++)
     {
         basis[ii] = LA::basis(ii, shape);
     }
 
     double k = 0.0;
-    for (size_t ii = 0; ii < dim; ii++)
+    for (int ii = 0; ii < dim; ii++)
     {
         k += commutator(a, commutator(b, basis[ii])).get_vector()[ii];
     }
@@ -55,20 +54,20 @@ Eigen::MatrixXd Killingform(const LA& g)
     * TODO: Return as glr type?
     */
 
-    const size_t dim = g.get_dimension();
-    const size_t shape = g.get_shape();
+    const int dim = g.get_dimension();
+    const int shape = g.get_shape();
     std::vector<LA> basis = std::vector<LA>(dim);
 
-    for (size_t ii = 0; ii < dim; ii++)
+    for (int ii = 0; ii < dim; ii++)
     {
         basis[ii] = LA::basis(ii, shape);
     }
 
     Eigen::MatrixXd K = Eigen::MatrixXd::Zero(dim, dim);
 
-    for (size_t ii = 0; ii < dim; ii++)
+    for (int ii = 0; ii < dim; ii++)
     {
-        for (size_t jj = 0; jj < dim; jj++)
+        for (int jj = 0; jj < dim; jj++)
         {
             K(ii, jj) = Killing(basis[ii], basis[jj]);
         }
